+++ include/tracey/ray/hittable_list.hpp
@@ -16,12 +16,12 @@ class HittableList : public Hittable{
     void Clear() {objects_.clear();}
     void add(shared_ptr<Hittable> obj) {objects_.push_back(obj);}
 
-    virtual bool Hit(const Tracey::Ray& r, float min, float max, HitRecord& hit) const override;
+    virtual bool Hit(const Tracey::Ray& r, double min, double max, HitRecord& hit) const override;
     
     std::vector<shared_ptr<Hittable>> objects_;
 };
 
-bool HittableList::Hit(const Tracey::Ray& r, float min, float max, HitRecord& hit) const{
+bool HittableList::Hit(const Tracey::Ray& r, double min, double max, HitRecord& hit) const{
     HitRecord temp;
     bool hit_anything=false;
     auto closest_so_far=max;
