+++ include/tracey/ray/sphere.hpp
@@ -7,18 +7,18 @@ namespace Tracey{
 class sphere : public Hittable {
     public:
         sphere() {}
-        sphere(Vector3f cen, float r) : center(cen), radius(r) {};
+        sphere(Vector3d cen, double r) : center(cen), radius(r) {};
 
         virtual bool Hit(
-            const Ray& r, float t_min, float t_max, HitRecord& rec) const override;
+            const Ray& r, double t_min, double t_max, HitRecord& rec) const override;
 
     public:
-        Vector3f center;
-        float radius;
+        Vector3d center;
+        double radius;
 };
 
-bool sphere::Hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const {
-    Vector3f oc = r.get_origin() - center;
+bool sphere::Hit(const Ray& r, double t_min, double t_max, HitRecord& rec) const {
+    Vector3d oc = r.get_origin() - center;
     auto a = r.get_direction().squaredNorm();
     auto half_b = oc.dot(r.get_direction());
     auto c = oc.squaredNorm() - radius*radius;
@@ -37,7 +37,7 @@ bool sphere::Hit(const Ray& r, float t_min, float t_max, HitRecord& rec) const {
 
     rec.t = root;
     rec.point = r.at(rec.t);
-    Vector3f outward_normal = (rec.point - center) / radius;
+    Vector3d outward_normal = (rec.point - center) / radius;
     rec.set_face_normal(r, outward_normal);
     //rec.normal = (rec.point - center) / radius;
